Add the following parameters that determines if the associated file is managed * $manage_epilog * $manage_prolog * $manage_task_epilog * $manage_task_prolog
If the value for $epilog, $prolog, etc, is not defined, the resource is still unmanaged.  This allows the path of the files to be defined but the files themselves to be unmanaged.

Allow values for $epilog and $prolog to be a glob (e.g. /etc/slurm/scripts.d/*).  When a glob is used the parent directory (e.g. /etc/slurm/scripts.d) is recursively managed and purged of unmanaged files.

Move $log_dir, $pid_dir and $shared_state_dir File resources to slurm::common::setup - these 3 directories will be managed if $controller, $slurmdbd or $node parameters are true.  The ownership is $slurm_user/$slurm_user_group for controller/slurmdbd and $slurmd_user/$slurmd_user_group for node.