Added 'acct_gather_profile/influxdb' plugin
This update introduces the installation of the plugin
'acct_gather_profile/influxdb'

The update is following the same way is with the HDF5 plugin, by
setting up 'None' as the default data type to be collect.
'profile_influxdb_username' and 'profile_influxdb_password' are
optional. The rest of parameters should be defined in order to make
this plugin to work properly.