Fix 'gres_types' in slurm.conf.erb
'gres_types' parameter is an 'Optional[Array[String]]', should contain
and Array of Strings or be 'undef'. However, template 'slurm.conf.erb'
does not reflect this and when gres_types is defined puppet crashes.

This update collects the string and applies a join separated by comma,
solving the mentioned problem.